+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.16
+
+- Editing a headline, when inserting or renaming a node, can now be interrupted by most other keyboard shortcuts and commands.
+- Fixed minibuffer history so that each opened Leo document has its own minibuffer command history: Requires Leo 6.7.5.
+- Implemented more of the original Leo's easter eggs: Pressing 'm' in the nav input launches the 'marked-list' command.
+- Added automatic Leo installation path finding.
+
 ## 1.0.15
 
 - Added description under command label in minibuffer-history's user input box.